pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "wagtailmeili"
description = "A Meilisearch backend for Wagtail CMS"
authors = [{name="Maxime Decooman", email="maxime@softquantum.com"},]
maintainers = [{name="Maxime Decooman", email="maxime@softquantum.com"},]
keywords = ["wagtail", "django", "search", "meilisearch", "backend"]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Wagtail",
    "Framework :: Wagtail :: 5",
    "Framework :: Wagtail :: 6",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]
dynamic = ["version"]
requires-python = ">=3.10"
dependencies = [
  "wagtail>=5.2",
  "meilisearch>=0.29",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-django",
    "ruff",
    "ruff-lsp",
    "python-decouple",
]

[project.urls]
Home = "https://github.com/softquantum/wagtailmeili"
Issues = "https://github.com/softquantum/wagtailmeili/issues"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "wagtailmeili.testapp.settings"
django_find_project = false
python_files = ['tests_*.py', 'test_*.py', '*_tests.py', '*_test.py']
filterwarnings = [
     'ignore::pytest.PytestAssertRewriteWarning:_pytest.config',
     'ignore::django.utils.deprecation.RemovedInDjango60Warning',
]
pythonpath = ["src"]
addopts = "--cov=wagtailmeili --cov-report=xml --cov-report=term-missing --reuse-db --no-cov-on-fail"

[tool.flit.sdist]
include = ["LICENSE", "README.md", "src/"]
exclude = [
    "src/wagtailmeili/tests/",
    "src/wagtailmeili/testapp/",
    ".github/",
    ".git/",
    ".gitignore",
    ".coverage",
    "coverage.xml",
    "*.pyc",
    "__pycache__/",
    "Makefile",
    "scripts",
    "docs",
    "*.db",
    "*.js",
    "*.json",
    "*.ini",
    "*.sqlite3",
    "*.yml",
    "*.yaml",
    "testmanage.py",
]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/testapp/*",
]